modified:   api/src/main/java/com/revature/controller/UserController.java 	modified:   api/src/main/java/com/revature/repository/UserRepository.java 	modified:   api/src/main/java/com/revature/service/UserService.java 	modified:   api/src/main/resources/application.properties 	modified:   api/src/main/resources/data.sql 	modified:   api/src/main/resources/schema.sql

+++ api/src/main/java/com/revature/service/UserService.java
@@ -35,7 +35,12 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteById(Long userId) {
-        userRepository.deleteById(userId);
+    public boolean deleteById(Long userId) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user != null) {
+            userRepository.deleteById(userId);
+            return true;
+        }
+        return false;
     }
 }
